+++ packages/terra-functional-testing/src/terra-cli/wdio/test-runner.js
@@ -41,6 +41,8 @@ class TestRunner {
    * @param {string} options.gridUrl - The remote selenium grid address.
    * @param {boolean} options.keepAliveSeleniumDockerService - Determines to keep the selenium docker service running upon test completion.
    * @param {string} options.locale - A language locale for the test run.
+   * @param {number} options.seleniumServicePort - The port mapping for the selenium service or the external selenium grid.
+   * @param {string} options.seleniumServiceUrl - The address for the selenium service.
    * @param {string} options.site - A file path to a static directory of assets. When defined, an express server will launch to serve the assets and disable running webpack..
    * @param {array} options.spec - A list of spec file paths.
    * @param {array} options.suite - Overrides specs and runs only the defined suites.
@@ -82,6 +84,8 @@ class TestRunner {
    * @param {string} options.gridUrl - The remote selenium grid address.
    * @param {boolean} options.keepAliveSeleniumDockerService - Determines to keep the selenium docker service running upon test completion.
    * @param {string} options.locales - A list of language locales for the test run.
+   * @param {number} options.seleniumServicePort - The port mapping for the selenium service or the external selenium grid.
+   * @param {string} options.seleniumServiceUrl - The address for the selenium service.
    * @param {string} options.site - A file path to a static directory of assets. When defined, an express server will launch to serve the assets and disable running webpack..
    * @param {array} options.spec - A list of spec file paths.
    * @param {array} options.suite - Overrides specs and runs only the defined suites.