+++ packages/terra-functional-testing/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Added
+  * Added seleniumServiceUrl parameter for locating selenium service.
+  * Added port parameter for specifying port mapping for the selenium service or the external selenium grid.
+
 ## 2.1.0 - (July 26, 2021)
 
 * Fixed